structs, gen: continue migration to signed types
for move_array_t
increase type width to 64 for most struct members

+++ src/search.c
@@ -130,7 +130,7 @@ int32_t quiescence(int32_t alpha, int32_t beta) {
    alpha = max(alpha, stand);
 
    move_array_t moves = sort_moves(generate_captures(state.side));
-   for (uint32_t i = 0; i != moves.count; ++i) {
+   for (int64_t i = 0; i != moves.count; ++i) {
       advance(moves.data[i]);
       tree_node_entry(alpha, beta);
       int32_t score = -quiescence(-beta, -alpha);
